requirements.txt
fastapi
uvicorn
strawberry-graphql
sqlalchemy
python-dotenv
redis